requirements.txt
python-dotenv==1.0.0
python_dateutil==2.8.2
requests==2.28.2
suds~=1.1.2
pyinstaller==5.10.0
pysimplegui~=4.60.4
dbfread~=2.0.7
platformdirs~=3.2.0
fuzzywuzzy~=0.18.0
chardet~=5.1.0
pywin32
despatchbay~=1.0.0